Cambio de tipo de peso de int a double en la clase Arista

+++ graph/Arista.java
@@ -4,6 +4,6 @@ public class Arista<E> {
 
     //Atributos
     private Vertice<E> referenciaDestino; // Vértice de destino de la arista
-    private int peso; // Peso de la arista 
+    private double peso; // Peso de la arista 
 
 }
